+++ internal/crlutil/crl_extensions.go
@@ -1,4 +1,4 @@
-package crl
+package crlutil
 
 import (
 	"bytes"
